Enhance Meal module, Dashboard, and Auth with new features

+++ src/app/modules/globalReviews/globalReviews.controller.ts
@@ -101,7 +101,7 @@ export const GlobalReviewsController = {
 
             const id = String(req.params.id);
 
-            // if (!req.user) throw new Error("User not found");
+            if (!req.user) throw new Error("User not found");
             const deletedReview = await GlobalReviewsService.deleteGlobalReview(id,req.user);
             
             res.status(200).send({
